requirements.txt
streamlit
tensorflow
transformers
opencv-python
pillow
numpy